Data container for MRexperiment objects

Used to serve metagenomic data (used in e.g., icicle plots and heatmaps). Wraps MRexperiment-class objects.

Methods

get_default_chart_type()

Get name of default chart type for this data type

get_measurements()

Get description of measurements served by this object

toMySQLDb(con, colLabel = NULL)

Save an MRexperiment object to a MySQL database.

toNEO4JDb(graph)

Save an MRexperiment object to a Neo4j Graph database.

toNEO4JDbHTTP(batch_url, neo4juser, neo4jpass, datasource, colLabel = NULL)

Save an MRexperiment object to a Neo4j Graph database.

toNEO4jDump(filePath = "~/", fileName = "neo4j_dump.cypher")

Save an MRexperiment object to a Neo4j Graph database. set the location of the output file with fileName and filePath.

update(new_object, send_request = TRUE)

Update underlying data object with new object