metavizr settings

Default settings for the various plotting functions in metavizr.

metavizControl(aggregateAtDepth = 3, aggregateFun = function(x) colSums(x),
  valuesAnnotationFuns = NULL, maxDepth = 4, maxHistory = 3,
  maxValue = NULL, minValue = NULL, title = "", n = 10000,
  rankFun = stats::sd, norm = TRUE, log = FALSE,
  featureSelection = NULL)

Arguments

aggregateAtDepth

Level of the tree to aggregate counts at by default.

aggregateFun

Function to aggregate counts by at the aggregateAtDepth level.

valuesAnnotationFuns

Function for error bars.

maxDepth

Level of the tree to display by default in icicle view.

maxHistory

Value for caching.

maxValue

Maximum value to display.

minValue

Minimum value to display.

title

title.

n

Number of OTUs to include in ranking.

rankFun

Ranking function - single vector function.

norm

Normalize MRexperiment object.

log

Log tranformation of MRexperiment object.

Value

List of setting parameters.

Examples

settings = metavizControl()